Moved MailStepExecutor into MailStep
