+++ support/src/main/java/org/jenkinsci/plugins/workflow/support/steps/mail/MailStep.java
@@ -23,11 +23,24 @@
  */
 package org.jenkinsci.plugins.workflow.support.steps.mail;
 
+import com.google.inject.Inject;
 import hudson.Extension;
+import hudson.model.TaskListener;
+import jenkins.plugins.mailer.tasks.MimeMessageBuilder;
+import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepImpl;
+import org.jenkinsci.plugins.workflow.steps.AbstractSynchronousStepExecution;
+import org.jenkinsci.plugins.workflow.steps.StepContextParameter;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
+
+import javax.mail.Address;
+import javax.mail.Message;
+import javax.mail.MessagingException;
+import javax.mail.Transport;
+import javax.mail.internet.MimeMessage;
+import java.io.UnsupportedEncodingException;
 
 /**
  * Simple email sender step.
@@ -83,6 +96,54 @@ public class MailStep extends AbstractStepImpl {
 
         @Override public String getDisplayName() {
             return "Mail";
+        }
+    }
+
+    /**
+     * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
+     */
+    public static class MailStepExecution extends AbstractSynchronousStepExecution {
+
+        private static final long serialVersionUID = 1L;
+
+        @Inject
+        private transient MailStep step;
+
+        @StepContextParameter
+        private transient TaskListener listener;
+
+        @Override
+        protected Object run() throws Exception {
+            MimeMessage mimeMessage = buildMimeMessage();
+            Transport.send(mimeMessage);
+            return null;
+        }
+
+        private MimeMessage buildMimeMessage() throws UnsupportedEncodingException, MessagingException {
+            if (StringUtils.isBlank(step.subject) || StringUtils.isBlank(step.body)) {
+                throw new IllegalArgumentException("Email not sent. All mandatory properties must be supplied ('subject', 'body').");
+            }
+
+            MimeMessage message = new MimeMessageBuilder()
+                    .setListener(listener)
+                    .setSubject(step.subject)
+                    .setBody(step.body)
+                    .setCharset(step.charset)
+                    .setMimeType(step.mimeType)
+                    .setFrom(step.from)
+                    .addRecipients(step.to, Message.RecipientType.TO)
+                    .addRecipients(step.cc, Message.RecipientType.CC)
+                    .addRecipients(step.bcc, Message.RecipientType.BCC)
+                    .setReplyTo(step.replyTo)
+                    .setDefaultSuffix(step.defaultSuffix)
+                    .buildMimeMessage();
+
+            Address[] allRecipients = message.getAllRecipients();
+            if (allRecipients == null || allRecipients.length == 0) {
+                throw new IllegalArgumentException("Email not sent. No recipients of any kind specified ('to', 'cc', 'bcc').");
+            }
+
+            return message;
         }
     }
 }
